% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian_estimation.r
\name{compute_est}
\alias{compute_est}
\title{compute network topology estimates for Gaussian source estimate}
\usage{
compute_est(graph, mean = 1, sigma = 1)
}
\arguments{
\item{graph}{igraph object specifying the underlying network graph}

\item{mean}{integer specifying normalizing constant for deterministic delay, default \code{mean=1}}

\item{sigma}{integer specifying normalizing constant for delay covariance, default \code{sigma=1}}
}
\value{
\code{compute_est()} returns a list of 
\itemize{
 \item \code{mu} matrix, estimate for deterministic delay, dimension (K-1)xN
 \item \code{Lambda} matrix, estimate for delay covariance, dimension (K-1)x(K-1),
} where K is the number of sensors and N the number of network nodes.
}
\description{
compute network topology estimates of deterministic delay and delay covariance for Gaussian source estimator
}
\examples{
# define network
set.seed(160716)
g <- igraph::sample_pa(30, directed=FALSE)
# partition network
g <- partition_network(graph = g, obs = c(1,12, 27))
# compute estimates
compute_est(graph = g)

}
