% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian_estimation.r
\name{couple_pair}
\alias{couple_pair}
\title{procedure for probability estimates for multi-class comparisons by pairwise coupling (Wu et al, 2004)}
\usage{
couple_pair(p2)
}
\arguments{
\item{p2}{symmetric matrix of pairwise probabilities}
}
\value{
numeric vector with estimated probability P(y=i|x) for each class
}
\description{
procedure for probability estimates for multi-class comparisons by pairwise coupling (Wu et al, 2004)
}
\details{
The implementation is an adapted version of \code{kernlab::minpair} (see references).
}
\references{
\itemize{
  \item Wu, Ting-Fan, Chih-Jen Lin, and Ruby C. Weng (2004). Probability estimates for multi-class classification by pairwise coupling. Journal of Machine Learning Research. 975-1005.
  \item Alexandros Karatzoglou, Alex Smola, Kurt Hornik, Achim Zeileis (2004). kernlab - An S4 Package for Kernel Methods in R. Journal of Statistical Software 11(9), 1-20. URL http://www.jstatsoft.org/v11/i09/
}
}
