% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{envirPara}
\alias{envirPara}
\title{\code{envirPara} Environment parameters for KwaZulu-Natal}
\description{
\code{envirPara} Environment parameters for KwaZulu-Natal
}
\details{
\code{envirPara} Environment parameters for KwaZulu-Natal
}
\examples{
library(NetOrigin)
data(envirPara)
y0 <- initial_condition_SIB_SS(popu, sigma, mu_B, theta, c(428))
}
\keyword{data}
