% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpd_gaussian.r
\name{hpd-methods}
\alias{hpd-methods}
\alias{print.hpd}
\alias{summary.hpd}
\alias{plot.hpd}
\title{methods for highest posterior density region objects of class \code{hpd}}
\usage{
\method{print}{hpd}(x, ...)

\method{summary}{hpd}(object, x = object, ...)

\method{plot}{hpd}(x, y = NULL, graph, ...)
}
\arguments{
\item{x}{object of class \code{\link{hpd}}}

\item{...}{further parameters to be passed to \code{\link{plot.igraph}}}

\item{object}{object of class \code{\link{hpd}}; passed to \code{x}}

\item{y}{\code{NULL}}

\item{graph}{\code{\link{igraph}} object specifying the underlying network graph}
}
\description{
\code{print} produces an output for objects of class \code{hpd}.
}
\seealso{
\code{\link{hpd}} \code{\link{hpd.origin}} \code{\link{origin-methods}}
}
