% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian_estimation.r
\name{partition_network}
\alias{partition_network}
\title{partitioning network for Gaussian source estimation}
\usage{
partition_network(graph, obs)
}
\arguments{
\item{graph}{igraph object, network graph}

\item{obs}{integer, observer nodes, while first node will be the reference node}
}
\value{
igraph object, with logical graph attributes \code{'obs'}, \code{'ref'}, \code{'Rset'}, \code{'Uset'}, and \code{'Lset'}. The node attribute \code{'set'} summarizes to partioning set assignment: 
\itemize{
\item \code{'O'} observer node
\item \code{'R'} resolvable node
\item \code{'L'} unresolvable node, adjacent to one observer
\item \code{'U'} unresolvable node, adjacent to a resolvable node from set R
}
}
\description{
partitioning network graph into sets of observers, resolvable sets, and unresolvable sets L and U.
}
\examples{
# define network
set.seed(160716)
g <- igraph::sample_pa(30, directed=FALSE)
#partition network
g <- partition_network(graph = g, obs = c(1,12, 27))
plot_partition(g, vertex.size=15, edge.width=2)

}
\seealso{
plot_partition
}
