<<<<<<< HEAD
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/origin_helper.r
\name{performance}
\alias{performance}
\title{generic method for performance evaluation}
\usage{
performance(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{further arguments}
}
\description{
generic method for performance evaluation
}
\seealso{
\code{\link{origin-methods}} \code{\link{plot_performance}}
}
=======
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/origin_helper.r
\name{performance}
\alias{performance}
\title{generic method for performance evaluation for objects of class \code{origin}}
\usage{
performance(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{further arguments}
}
\value{
none, there are no applicable methods available for other objects
}
\description{
generic method for performance evaluation for objects of class \code{origin}
}
\seealso{
\code{\link{origin-methods}} \code{\link{plot_performance}}
}
>>>>>>> 62387598c0dde62f2a36b8be7ec984356b464310
