% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/origin_helper.r
\name{plot}
\alias{plot}
\title{generic method for plots}
\usage{
plot(x, y, ...)
}
\arguments{
\item{x}{object}

\item{y}{object}

\item{...}{further arguments}
}
\description{
generic method for plots
}
\seealso{
\code{\link{plot_partition}} \code{\link{plot_performance}} \code{\link{plot_ptn}} \code{\link{origin-methods}} \code{\link{hpd-methods}}
}
