% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian_estimation.r
\name{plot_partition}
\alias{plot_partition}
\title{plot partitioned network}
\usage{
plot_partition(graph, ...)
}
\arguments{
\item{graph}{igraph object specifying the network graph with attribute \code{'set'} defining the partion}

\item{...}{further parameters to be passed to \code{\link{plot.igraph}}}
}
\description{
plot partitioned network with coloring according to sets of observers, resolvable sets, and unresolvable set L and U.
}
\examples{
# define network
set.seed(160716)
g <- igraph::sample_pa(30, directed=FALSE)
#partition network
g <- partition_network(graph = g, obs = c(1,12, 27))
plot_partition(g, vertex.size=15, edge.width=2)

}
\seealso{
partition_network plot_ptn
}
