% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/origin_helper.r
\name{print}
\alias{print}
\title{generic methods for printed output}
\usage{
print(x)
}
\arguments{
\item{x}{object}

\item{...}{further arguments}
}
\description{
generic methods for printed output
}
\seealso{
\code{\link{print.origin}}, \code{\link{origin-methods}}, \code{\link{print.robustness}}, \code{\link{print.hpd}}, \code{\link{hpd-methods}}
}
