% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian_estimation.r
\name{prior}
\alias{prior}
\title{compute prior distribution for Gaussian source estimator}
\usage{
prior(graph, type = c("equal", "random", "ring"), zeta = 1, center = NULL)
}
\arguments{
\item{graph}{igraph object specifying the underlying network graph}

\item{type}{character, specifying the prior type to be computed (see details)}

\item{zeta}{numeric, scaling parameter for informative priors}

\item{center}{integer, identifying the center of informative priors}
}
\value{
numeric vector of length \code{vcount(graph)}.
}
\description{
compute prior distribution for Gaussian source estimator
}
\details{
Different prior specifications are implemented: 
\itemize{
   \item \code{type='equal'} assigns equal priors to all nodes, which equals in the Gaussian source estimator without priors (Pinto et al., 2012)
   \item \code{type='random'} is an non-informative prior with \deqn{\pi \sim N(\mu=1/N, \sigma=1/N^4)}
   \item \code{type='ring'} is an informative prior with exponential decay according to the shortest path distance from the \code{center} node and scaling parameter \code{zeta}, i.e. \deqn{\pi_i(s) = \exp(-\zeta \cdot d(s,i))}
} All priors are normalized, so that \deqn{\sum_{i=1}^N \pi_i = 1}
}
\examples{
# define network
set.seed(160716)
g <- igraph::sample_pa(30, directed=FALSE)
# compute prior
prior(g, type='equal')
prior(g, type='random')
prior(g, type='ring', zeta=1, center=8)

}
\references{
\itemize{
  \item Pinto, P. C., P. Thiran, and M. Vetterli (2012). Locating the source of diffusion in large-scale networks. Physical Review Letters 109 (6).
}
}
\seealso{
\code{\link{origin_gaussian}}
}
