% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handling.r
\name{read_DB_data}
\alias{read_DB_data}
\title{Reads a data file as provided by 'Deutsche Bahn' (for internal use).}
\usage{
read_DB_data(file)
}
\arguments{
\item{file}{character with path and file name containing the variables for  'stationID', 'date', 'hour', 'minutes', and 'delay'}
}
\value{
\code{data.frame} with variables \code{'node'}, \code{'time'}, \code{'delay'}
}
\description{
Reads a data file as provided by 'Deutsche Bahn' (for internal use).
}
\seealso{
Other data_handling: 
\code{\link{aggr_data}()}
}
\concept{data_handling}
