% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustness.r
\name{robustness-methods}
\alias{plot.robustness}
\alias{print.robustness}
\alias{robustness-methods}
\alias{summary.robustness}
\title{methods for robustness estimation objects of class \code{robustness}}
\usage{
\method{print}{robustness}(x, ...)

\method{summary}{robustness}(object, x = object, ...)

\method{plot}{robustness}(x, y = NULL, add = FALSE, ...)
}
\arguments{
\item{x}{\code{data.frame} obtained by \code{\link{robustness}}, robustness estimation object for source estimation from function \code{\link{robustness}}}

\item{...}{further arguments passed to the default \code{print} method}

\item{object}{object of class \code{\link{origin}}, origin estimation object from function \code{origin_xxx}; passed to \code{x}}

\item{y}{not used; default \code{NULL}}

\item{add}{logical specifying whether this should be added to another robustness plot}
}
\description{
\code{print} produces an output for objects of class \code{robustness}

\code{summary} produces an object summary for objects of class \code{robustness}

\code{plot} produces a time series plot of the \code{robustness} estimate object
}
\seealso{
\code{\link{robustness}}
}

