% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIB_SS.R
\name{SIB_SS}
\alias{SIB_SS}
\title{Stochastic SIB model for infected cases simulation}
\usage{
SIB_SS(
  mu,
  beta,
  rho,
  sigma,
  gamma,
  alpha,
  mu_B,
  m = 0.3,
  theta,
  nnodes,
  POP_node,
  fluxes,
  time_sim,
  y0
)
}
\arguments{
\item{mu}{population natality and mortality rate (day^-1)}

\item{beta}{contact rate}

\item{rho}{immunity loss rate (day^-1)}

\item{sigma}{symptomatic ratio, i.e., fraction of infected people that develop symptoms and are infective. 
(The remaining fraction enters directly the recovered compartment.)}

\item{gamma}{rate at which people recover from cholera (day^-1)}

\item{alpha}{cholera induced mortality rate (day^-1)}

\item{mu_B}{death rate of V.cholerae in the aquatic environment (day^-1)}

\item{m}{parameter for infection force, default value is 0.3}

\item{theta}{contamination rate}

\item{nnodes}{number of nodes/cities}

\item{fluxes}{matrix, number of nodes x number of nodes 
where each row contains the probabilities a person travels from the given city (by Row Index) to another city (by Column Index).}

\item{time_sim}{time steps for simulation, e.g., seq(0, 100, 0.1)}

\item{y0}{initial condition for SIB_SS, output of 'initial_condition_SIB_SS'}

\item{POP_nodes}{vector, length represents number of cities/nodes; vector represents
population at each node}
}
\value{
a matrix, nnodes x number of time steps, representing number of new cases at each node, each time step
}
\description{
\code{SIB_SS} Stochastic SIB model for infected cases simulation
}
\examples{
library(NetOrigin)
data(envirPara)
y0 <- initial_condition_SIB_SS(popu, sigma, mu_B, theta, c(428))
time_sim=seq(0, 100, by=0.1)
simu.list = SIB_SS(mu = mu, beta = beta, rho = rho, sigma = sigma, gamma = gamma,
                   alpha = alpha, mu_B = mu_B, theta = theta, nnodes = length(POP_node), POP_node = popu,
                   fluxes = humanmob.mass, time_sim = time_sim, y0 = y0)
}
