% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/origin_helper.r
\name{summary}
\alias{summary}
\title{generic methods for summaries}
\usage{
summary(object, ...)
}
\arguments{
\item{object}{object}

\item{...}{further arguments}
}
\description{
generic methods for summaries
}
\seealso{
\code{\link{summary.origin}}, \code{\link{origin-methods}}, \code{\link{summary.robustness}}, \code{\link{summary.hpd}}, \code{\link{hpd-methods}}
}
