% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/origin_helper.r
\name{performance}
\alias{performance}
\title{generic method for performance evaluation}
\usage{
performance(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{further arguments}
}
\description{
generic method for performance evaluation
}
\seealso{
\code{\link{origin-methods}} \code{\link{plot_performance}}
}
