% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/origin_helper.r
\name{plot_performance}
\alias{plot_performance}
\title{A plot method combining a time series of performance results.}
\usage{
plot_performance(
  x,
  var = "rank",
  add = FALSE,
  offset = NULL,
  log = FALSE,
  col = 1,
  ylim = NULL,
  text.padding = 0.9,
  ...
)
}
\arguments{
\item{x}{\code{data.frame} obtained by combined results from \code{\link{performance.origin}} with variables \code{X1} for time point, \code{start} for true origin, \code{est} for estimated origin, and performance variables}

\item{var}{character, variable to be plotted, \code{\link{performance.origin}} returns \code{rank}, \code{spj}, and \code{dist}, default is \code{'rank'}}

\item{add}{logical, should be added to another performance plot}

\item{offset}{\code{POSIXct}, starting time of spreading}

\item{log}{logical, should y-axis be logarithmized?}

\item{col}{numeric or character, color of lines}

\item{ylim}{numeric vector, range of y axis}

\item{text.padding}{a numeric value specifying the factor for the text position relative to the y values}

\item{...}{further graphical parameters passed to default \code{plot} function}
}
\value{
No return value
}
\description{
A plot method combining a time series of performance results.
}
